fix(template/common): change ESLint default config
Add the `"root": true` flag to the default ESLint config in the `package.json` for the common template.
If this flag is not set, ESLint will automatically look for them in the directory of the file to be linted, and in successive parent directories all the way up to the root directory of the filesystem (https://eslint.org/docs/user-guide/configuring#using-configuration-files-1). So if using your new package e.g. as a git submodule, the package's config can conflict with the parent project's config, causing ESLint to fail